requirements.txt
mlflow
torch
torchvision
Pillow
tb-nightly
future
numpy
pandas
opencv-python
matplotlib
coloredlogs
ruamel.yaml
tqdm